elt/requirements.txt
pandas
psycopg2-binary
SQLAlchemy
polars
adbc-driver-postgresql
pyarrow
python-dotenv
boto3
